+++ src/nuxt/types.ts
@@ -13,3 +13,6 @@ declare module '@nuxt/schema' {
     visualFeedback?: ModuleOptions;
   }
 }
+
+export type { ModuleOptions } from './module';
+export { default } from './module';